06-django-lab/01-django-app/06-emerce/requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-session-timeout==0.1.0
idna==2.10
Pillow==9.0.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
